Quitar classes menos human
