+++ practica_REFAC/src/Customer.java
@@ -1,0 +1,13 @@
+public class Customer {
+    private String firstName;
+    private String lastName;
+
+    public Customer(String firstName, String lastName) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+    }
+
+    public String getFullName() {
+        return firstName + " " + lastName;
+    }
+}
