+++ practica_REFAC/src/Order.java
@@ -1,0 +1,13 @@
+import java.util.List;
+
+public class Order {
+    private List<OrderLineItem> lineItems;
+    private double taxRate;
+
+    public Order(List<OrderLineItem> lineItems, double taxRate) {
+        this.lineItems = lineItems;
+        this.taxRate = taxRate;
+    }
+
+
+}
